xpay: add amnesia mode to forget hints after payment
This is useful in testing the performance of xpay with payment
failure feedback which is typically stored on the xpay layer of
askrene.  In dev_amnesia mode, hints are stored on an ephemeral
layer instead which is discarded before the next call.

Changelog-Added: xpay dev_amnesia flag: developer option to forget askrene hints after a payment attempt completes